optimize rerender on inputbox
